02_vectorize/requirements.txt
pandas
numpy
tqdm
sentence-transformers
ollama
PyYAML